{{cookiecutter.project_slug}}/backend/requirements/base.txt
django>=4.0, <5.0
psycopg2-binary>=2.9.3, <3.0
python-environ>=0.4.54, < 1.0

djangorestframework>=3.13.1, <4.0
django-cors-headers>=3.10.1, <4.0
djoser>=2.1.0, <3.0
djangorestframework-simplejwt>=4.3.0, <5.0

whitenoise>=5.3.0, <6.0